ISPN-7007 Added org.infinispan.remote module
