+++ integrationtests/wildfly-modules/src/test/java/org/infinispan/test/integration/as/client/Person.java
@@ -1,0 +1,16 @@
+package org.infinispan.test.integration.as.client;
+
+import org.infinispan.protostream.annotations.ProtoField;
+
+public class Person {
+
+   @ProtoField(number = 1)
+   public String name;
+
+   public Person(String name) {
+      this.name = name;
+   }
+
+   public Person() {
+   }
+}
