+++ wildfly-modules/build.xml
@@ -199,6 +199,8 @@
         <module-def name="org.hibernate.hql.parser" slot="${infinispan.slot}">
             <maven-resource group="org.hibernate.hql" artifact="hibernate-hql-parser" />
         </module-def>
+
+        <module-def name="org.infinispan.remote" slot="${infinispan.slot}"/>
     </target>
 
     <target name="all" depends="clean, prepare-hibernate-search-module, copy-files, modules"/>
