+++ wildfly-modules/src/main/resources/org/infinispan/remote/main/module.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<module xmlns="urn:jboss:module:1.1" name="org.infinispan.remote" slot="${slot}">
+    <dependencies>
+        <module name="javax.api"/>
+        <module name="org.infinispan.client.hotrod" slot="${slot}" export="true"/>
+        <module name="org.infinispan.commons" slot="${slot}" export="true"/>
+        <module name="org.infinispan.protostream" slot="${slot}" export="true"/>
+        <module name="org.infinispan.query.dsl" slot="${slot}" export="true"/>
+        <module name="org.infinispan.query.remote.client" slot="${slot}" export="true"/>
+        <module name="org.infinispan.cdi.remote" slot="${slot}" export="true"/>
+        <module name="org.jboss.logging"/>
+    </dependencies>
+</module>
